Add utility method for doing sample type template download with import aliases
