+++ packages/components/src/index.ts
@@ -345,6 +345,7 @@ import {
     getSampleStatus,
     getSampleStatusType,
     getSampleTypeTemplateUrl,
+    downloadSampleTypeTemplate,
     isSampleOperationPermitted,
     isSamplesSchema,
     SamplesManageButtonSections,
@@ -991,6 +992,7 @@ export {
     getSampleStatus,
     getSampleStatusType,
     getSampleTypeTemplateUrl,
+    downloadSampleTypeTemplate,
     DisableableMenuItem,
     SampleOperation,
     SampleStateType,
